Work in progress - EOD
Almost ready to take a bunch of GA graphs and data. However, I think there are some issues with the way the goal array and goal indices are defined and calculated. Need to think about whether there would be an issue if the pixels which ARE changed are not separated from the pixels whiche are NOT changed (initial beam above vs below the goal). Also need to figure out how the indices are calculated and used. Additional feature to select and input a custom threshold with both value (0-255) and percent (0-100) inputs